Update start to work w/ staging environment
* You should now use a `./.env` for the environment
* On a real server it should define `LAPIS_ENVIRONMENT`.

The reason for choosing `.env` is that there are a number of tools which use `.env` as the default file for local configurations. 
One of this is [Overmind](https://github.com/DarthSim/overmind) which basically just makes managing processes easier, since it can handle an app server + db and so on.