Add Hofstra University
ITEEA and Hofstra University are creating an NSF funded project to create an out-of-the-box computer science class entitled "The Beauty and Joy of Computing" and we are using Snap as a part of the project. We are implementing a version of NCSU's snap lessons, but are hosting on Hofstra's servers. Please allow the following:

http://techlit.org/
http://snap.techlit.org/

Thank you!